Solution to Input and Print an Integer

+++ Input and Print an Integer@@6788d5012db25efec946f939/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float a;
+    int a;
     printf("" , a);
-    scanf("%f", a);
+    scanf("%d", a);
     return 0;
 }
